Make PostAsync() overrideable
In ForgeOperationsQuery, `public Task<JToken> PostAsync(string contents)`
is being overridden, but `PostAsync` in RpcPost is not marked as virtual.

This commit assumes that all `PostAsync` methods in RpcPost are allowed
to be overridden even though only one is currently being overridden.

And it marks that one case with `override`.